pyproject.toml
[project]
name = "shifts-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pip>=25.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "tzdata>=2.5.2",
]

